+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -7,7 +7,7 @@ object LibExt {
     var libVersion = ""
 
     const val gdxVersion = "1.12.1"
-    const val teaVMVersion = "0.10.2"
+    const val teaVMVersion = "0.11.0"
 
     const val gdxImGuiVersion = "-SNAPSHOT"
     const val gdxMultiViewVersion = "-SNAPSHOT"