+++ examples/freetype/teavm/src/main/java/com/github/xpenatan/gdx/examples/teavm/FreetypeTestLauncher.java
@@ -13,7 +13,7 @@ public class FreetypeTestLauncher {
         config.showDownloadLogs = true;
 
         config.preloadListener = assetLoader -> {
-            assetLoader.loadScript("freetype.js", null);
+            assetLoader.loadScript("freetype.js");
         };
 
         new TeaApplication(new FreetypeDemo(), config);