+++ examples/gdx-tests/teavm/src/main/java/com/github/xpenatan/gdx/examples/teavm/GdxTestLauncher.java
@@ -15,7 +15,7 @@ public class GdxTestLauncher {
         config.useGL30 = true;
 
         config.preloadListener = assetLoader -> {
-            assetLoader.loadScript("freetype.js", null);
+            assetLoader.loadScript("freetype.js");
         };
 
         new TeaApplication(new ImGuiGame(), config);