+++ CHANGES.md
@@ -1,7 +1,10 @@
 [-SNAPSHOT]
+
+[1.0.5]
 - Add Config Asset preloadListener
 - Update AssetLoader and AssetDownloader
 - Add AssetInstance to obtain AssetLoader or AssetDownloader
+- Update Freetype emulation to fix build errors and script loading solution
 
 [1.0.4]
 - Fix music id